drop SATA parts of server_(sata)_multi (goes with D730)
Summary:
D730 drops the SATA parts of the server_sata_multi test and
renames it to server_multi, see D730 and T691 for the reason.
This adjusts the wiki reporting config for the new test name
and removes the SATA test instance.

Test Plan:
After running tests with D730, generate results
and make sure the appropriate ResTuples show up.

Reviewers: jskladan, garretraziel

Reviewed By: garretraziel

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D731