Pungi 4 conversion: no more fedfind
This is a major revision of the scheduler with the intent of
scheduling jobs based on Pungi 4-provided metadata, rather
than fedfind information (which is kind of reverse-engineered
compose metadata).

We throw out all fedfind usage. We also throw out all support
for compose waiting. We *also* throw out support for running
on the 'current' validation compose. The idea is that we will
simply run against every compose, as it appears, thanks to
Pungi 4 sending fedmsg messages when composes complete: there
will be a fedmsg listener (it may be a new tool, or we may add
it to this codebase, or we may have taskotron do it) which
listens out for 'compose complete' messages, and calls the
scheduler. So the only way to schedule jobs now is to run the
scheduler (or import the module and call `jobs_from_compose()`),
and you have to provide the compose ID and the compose location,
both of which are in the fedmsg. Manual invocation is still
possible so long as you provide both those values.

The compose must exist, there is no point in implementing
'waiting' if we can rely on having a 'compose complete' signal
from fedmsg.

The scheduler now uses the Pungi `compose_id` as the openQA
BUILD, and derives the openQA FLAVOR from the Pungi `variant`,
`type` and `format`. There's no attempt to modify these values
to match the old fedfind-style values, I figured we should just
make a clean break and respect the Pungi values. The wiki
report code will have to translate between Pungi `compose_id`
and Wikitcms validation event versioning for now; I may look
at changing the Wikitcms validation event versioning later.
For now it should handle nightly events, I do not know yet how
the Pungi metadata for RCs will look so it doesn't handle those
yet (TCs may well no longer exist).

Pungi, unfortunately, doesn't treat 'Rawhide' as a release; it
synthesizes a release number for Rawhide nightlies. I think
this is wrong and I've reported it, but for now, we just set
VERSION to Rawhide if we see the string 'rawhide' in the
compose URI anywhere.

The way we specify which images we want to test is also changed
completely. I figured it'd be neat to follow the Pungi
`images.json` metadata layout; our list of 'wanted images' is
basically a subset of that file, and if all the properties of
any item in our subset match one of the images in the compose,
we count that as a match and test that image. There's one
special value, `score`, which doesn't exist in Pungi's metadata
and isn't used for image matching, it's instead used to pick
which image we'll use for the 'universal' tests (this seemed
like the neatest design for doing that).

We also dump the image download code and use `ISOURL` instead
of `ISO`, that's the feature I added to openQA a while back
which lets it download the ISO for testing. This is not actually
needed to work with Pungi 4, but since we're thinking of having
Taskotron do the scheduling, it seemed wise; we have to use
this 'server download' feature if we want the scheduler to run
on any box other than the openQA server itself.