make get_images 'private', add _get_compose_id
I don't think we really want to advertise get_images, so turn
it into _get_images. Also, fedmsg's design has a limitation
that means we can only really get *either* compose_id *or*
location for a compose (not both) without ugly workarounds, so
instead of doing ugly workarounds, let's just assume we'll only
get location, and discover the compose_id from the metadata.
The only drawback of this is it adds another remote roundtrip,
but that's not really a big deal.