the scheduler moved wrt README.md and setup.py broke, fix it
Pushing without review as this is breaking ansible runs on the
prod instances and we need to fix them. Moving the scheduler
into a subdirectory broke this, as `read()` assumed the file
given was in the same directory. As it's used only for this,
just change the path in `read()`.