pungi4 images: fudge up 'payload', separate match dict
As we get bigger pungi4 composes I start to see the unfortunate
bits in the metadata...

First up, on second thought I don't like just smooshing two
different types of information together in the 'WANTED' image
dicts, so let's keep all the keys used for image matching in
a dict called 'match'. Other items in each image dict will be
things used for configuring the scheduler's behaviour. Right
now this is only 'score', but we'll probably add other things
like this later, perhaps for cuing when we need to do a direct
kernel boot (though we haven't entirely thought that through
yet).

Second, I was expecting (hoping?) that each live image would be
its own variant, but alas, no. Seems that 'variants' are tied
to 'subdirectories', so our 'variants' are things like 'Spins'
and 'Labs' - fairly useless for image identification. Nothing
in pungi4 or productmd reliably provides an equivalent to the
fedfind 'payload' concept. So we have to sneak in some ugly
fedfind-lite code to derive a 'payload' property from the image
path. Without this we can't reasonably tell the KDE live image
from the LXDE live image or any other live image in the "Spins"
variant. I will talk to dgilmore and dmach about improving this,
but for F24 I think we're probably stuck with it.