actually, no, let's just hardcode the atomic stuff
This reverts commit 140a538489d753d4422de7ed6561d112bd9bd867
and replaces it with something else.

After a good night's sleep I decided I didn't really like it,
it's too clever-clever. Let's just be dumb and hardcode stuff.
We know the Two Week Atomic nightlies are the only non-Pungi
4 composes we're actually going to care about, and we know
precisely what image we want from them and where it will be.
So we can just have some stupid temporary code that knows to
go and find exactly that image and give it the properties we
want it to have. This is fine. This is fine. This is fine.