add a simple fedmsg listening job scheduler (John Dulaney)
Thanks to John for this. We want to have this done by taskotron
in the long term, but in the short term it seems releng will
switch to Pungi 4 composes very soon, and it's easier to write
a simple fedmsg listener to do scheduling for Pungi 4 composes
than to modify fedfind to support 'compose waiting' and so on
for them, or any other possible approach.