scheduler: restart the consumer when it crashes
Summary:
The fedmsg consumer crashed last night; apparently fedmsg
tail_messages() will crash by design if it can't retrieve the
CRL (for Security Purposes). Apparently the box's network went
down for a short time or something. This is obviously Bad -
we rely on the consumer running or else we'll miss scheduling
jobs for new composes. So let's have the unit file restart it
on failure. Just in case we write some dumb code that makes it
crash *all the time*, set RestartSec to 10 so it'll only try
and restart every ten seconds and won't kill the system or spam
the logs too badly.

Test Plan:
Start the consumer running, make it die somehow
(probably just killing it will do the trick), check it comes
back.

Reviewers: jskladan, garretraziel

Reviewed By: garretraziel

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D758