redo fedmsg consumer as a proper hub consumer
Summary:
We ran into issues with the check-compose consumer and the
openQA consumer colliding; it seems running two of the 'naive'
consumers (`tail_messages()`-based) on the same system is not
a good idea. This is the 'correct' way to build a production
consumer, according to fedmsg folks, and has various benefits
for Fedora infra deployments.

I've already re-done the relval and check-compose consumers in
this style, they seem to be working. Documentation for this
is at http://www.fedmsg.com/en/latest/consuming/ .

Test Plan:
Deploy the consumer as documented in README.md,
see if it works (you'll need to wait for a real compose, or
twiddle with the code a bit to listen for dev messages for
testing...I have a 'TestConsumer' model in the relval consumer
which might be worth adopting here.)

Reviewers: jskladan, garretraziel

Reviewed By: garretraziel

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D765