scheduler: make result reporting work again
committing without review as what's there will definitely not
work at all. this works...sorta...if you have latest git master
wikitcms...and, er, for Alpha 1 you need an ugly hack which
lives on my machine. but apart from that it works! this is the
correct code for here, anyway, the rest of the fun happens in
wikitcms and fedfind.