scheduler: use productmd 'subvariant' property
Summary:
productmd/releng folks kindly added a 'subvariant' property to
productmd image dicts, which fills the gap that we had to cover
by synthesizing 'payload'. This adjusts the scheduler to use it
and ditch the payload synthesizing stuff. Makes things a lot
cleaner! The only fiddle we have to do is to strip '_Base' from
the subvariant value if it's there; the subvariant for Cloud
disk images is 'Cloud_Base', but in Wikitcms result pages, we
always use 'Cloud' as the identifier...this actually shouldn't
matter at preset as we don't test any cloud images, but just in
case.

Test Plan:
Test that job scheduling and result reporting all work
exactly as before. The subvariant values for all images we test
should be identical to the old 'payload' values, so all the openQA
flavor names should be unchanged.

Reviewers: jskladan, garretraziel

Reviewed By: garretraziel

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D784