scheduler: test Workstation netinst by default too
Summary:
this should be all that's needed to test workstation netinst,
I think. The bits are already there in templates. Apparently
we still care about this image, so hey.

Test Plan:
Do a full test run for a compose where the Workstation
boot.isos completed, and check they get tested.

Reviewers: jskladan, garretraziel

Reviewed By: garretraziel

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D788