scheduler: add firewall test case config
Summary:
This goes along with D792. At first I was thinking of exciting
complicated ways to 'group' tests such that a whole group of
tests has to pass before the wiki result will be reported. Then
I realized that was unnecessary complication and we can just
split up the damn wiki test case. So this change refers to two
wiki test cases that don't actually exist yet; the idea is if
reviewers are OK with this, we'll split
QA:Testcase_kickstart_firewall into two in the matrices.

Test Plan:
Well it's a bit hard as the wiki isn't split; you
could fake up a Server result page in the staging wiki with the
test case split up and make sure it works, I guess. But it's a
pretty straightforward change, shouldn't really be dangerous.

Reviewers: jskladan, garretraziel

Reviewed By: garretraziel

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D793