allow passing extra params to ISO post, updates URL to CLI
Summary:
this commit adds a fairly generic mechanism for passing extra
params into the ISO post command when scheduling jobs, and an
arg to the `compose` sub-command which takes advantage of this
to let you specify an updates image URL which will cause the
new `GRUBADD` var to be specified when POSTing the jobs, thus
loading that updates image for all the tests. This should give
us a handy way to run a full openQA test run on any updates
image we like.

When extra parameters are specified, the BUILD value is changed
so this test run can be identified as not being a 'pure' one.
get_passed_testcases is tweaked to make sure we never report
results to the wiki for such tests.

This goes with D801 for the tests.

Test Plan:
With D801 applied to the tests, try something like
`fedora-openqa-schedule.py compose -f --updates http://dlehman.fedorapeople.org/updates/1322943.0.img https://kojipkgs.fedoraproject.org/compose/rawhide/Fedora-Rawhide-20160331.n.0/compose/`
and see what happens; it should add a `GRUBADD` var to all
tests which should cause that updates image to be loaded. The
updates image loading tests will fail, but all other tests
should now get further than they would without the updates
image (which fixes an early crasher bug). The build should be
the compose ID with -EXTRA appended, and if you try reporting
results to the wiki it should find no passed testcases, with
a debug log message. I will need to adjust the check-compose
consumer to ignore such builds (though we should never run
tests like this on the production openQA, I don't think, at
most we might do it on staging) and I'll also adjust the wiki
report consumer not to even bother trying to report results
for these composes, just to double the protection.

Reviewers: jskladan, garretraziel

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D802