bump the DVDs up in 'universal' scoring again
Summary:
this goes with D844. With that change we should be able to use
the DVDs for universal testing again, so let's do that: it
avoids transient mirror problems and also avoids the 'tests run
before mirror metadata syncs' problem.

Test Plan:
Run the scheduler on a compose with Server DVDs and
make sure it uses them for the universal tests.

Reviewers: jskladan, garretraziel

Reviewed By: garretraziel

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D845