add reporting to ResultsDB
This adds possibility to report results to ResultsDB. It
changes cli args a little bit (now you specify either `--wiki` or
`--resultsdb` or both if you want to submit your
results). It sets more variables in openQA during test scheduling,
namely subvariant and imagetype (we COULD retrieve those things
from BUILD, but it's safer this way) as well as job ID of ResultsDB
job (that had to be created during test scheduling).

In contrast to Wiki reporting, we want to report both passes and
fails to ResultsDB (we probably aren't going to use ResultsDB directly,
we are planning to write some service on top of it).

We are using shiny new convention of naming tests with dot separator,
so `QA:Testcase_Boot_default_install` becomes
`openqa.installation.boot_default_install`. We are also trying to put
links where appropriate, so job overview links to openQA's overview,
testcase links to testcase page on wiki and result links to openQA
job run.

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D925