rework wiki report 'conditions', for notification tests
Summary:
The big change here is a rework of how 'advanced' entries in
conf_test_suites TESTSUITES work. To report the notification
test results we need to allow for this logic: 'report a pass
for wiki test X only if both testsuite A and testsuite B
passed' - neither desktop_notifications_live passing on its
own nor desktop_notifications_postinstall passing on its own
fully covers any of the wiki test cases, but if *both* pass,
then two of the wiki test cases are covered.

I wanted a consistent and vaguely understandable syntax for
expressing this, so I incorporated a rework of the existing
complex case - 'report a pass only if specific openQA test
modules pass' - into this too. Thus the layout of the dicts
used in the 'complex' case is changed so it's less specific
to the 'module' case. The dicts now use test case names as
their keys and dicts of 'conditions' that must be satisfied
for a pass to be reported as the value for each key. An empty
'conditions' dict just means the same as the 'basic' format:
report this test case as passed if the job passed.

Otherwise, the dict can specify various different types of
conditions that must be satisfied *as well as* the job passing
before a wiki pass will be generated for that wiki test case.
Currently there are 'modules' and 'testsuites' conditions, but
with this design we could add more if needed. The 'modules'
value is an iterable of openQA job modules which must have
passed within the job for a wiki pass to be generated; the
'testsuites' value is an iterable of other test suite names,
each of which must have passed for the same openQA BUILD,
MACHINE and FLAVOR for a wiki pass to be generated.

We adjust the FreeIPA tests to the new design, add entries for
the notifications tests using the new design, and adjust
`report.get_passed_testcases` to do the work. We fix and update
some comments, and split the now-quite-long code for parsing
the TESTSUITES dicts into its own function, just to make things
a bit more readable.

We also adjust `checkwiki` for the changed design, enhance it
to handle the 'testsuites' condition and run some specific
checks for it, enhance it to check that all 'modules' named in
a condition are present in the openQA distri (as I realized
that's a check that was missing), and refactor it into separate
test functions, as the giant `main()` was getting unwieldy,
hard to follow, and increasingly likely to suffer some kind of
variable scope issue.

Test Plan:
Use checkwiki and mess around with conf_test_suites
a bit, generate results for a test run that has passed FreeIPA
and notifications tests and make sure you get appropriate
ResTups. I tested reporting the 20160927.n.0 results with this
code and you can see it worked:
https://fedoraproject.org/wiki/Test_Results:Fedora_25_Branched_20160927.n.0_Desktop
It'd probably be good to somehow force a case where one of the
notifications tests passed but the other failed, and check we
do *not* report a pass in this case (fake up the jobs or just
find a compose where that happened or whatever).

Reviewers: jskladan, garretraziel

Reviewed By: garretraziel

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D1018