use fedfind for flavor and imagetype construction
This is needed to work around a significant issue with the
metadata provided by Pungi:
https://pagure.io/pungi/issue/417
basically, the metadata as it comes out of Pungi cannot tell
the Workstation ostree installer image from the Workstation
network install image. As we actually test the network install
image, this is a major problem as we might wind up grabbing the
ostree installer image and testing it as if it were the network
install image.

To work around this, use the new fedfind `correct_image` helper
on the image dicts we read from the metadata; it was added
specifically to work around this problem (and provide a place
to work around similar problems in future). It will change the
type for ostree installer images to 'dvd-ostree'.

Pushing without review as this problem needs fixing immediately
to ensure we test the right images. Please let me know of any
issues or improvements we should make to this code.