Support testing the semi-official live-respins, more fedfind
Summary:
The point of this change is to support testing the 'respin'
composes that are built by volunteers and live in live-respins
in the alt tree on the mirrors. These are considered to be
'semi-official', and the maintainers asked if we could test
them. This depends on the latest fedfind git master ATM (I'll
cut a release soon).

To make this possible we kinda have to go back to using fedfind
more. I was hoping when I initially dumped the use of fedfind
that fedfind might actually go away soon, but for a bunch of
awful reasons that make me want to drink, this seems very
unlikely. These live-respins composes aren't built with Pungi 4
and don't have metadata, so we kinda have to go through fedfind
to understand them.

I think the changes to use fedfind actually make the scheduler
code a bit nicer, though. Especially it lets us use a less
stupidly-nested 'WANTED' structure (now a list not a dict).

I have tried this out on staging (that's why there were a bunch
of needle commits to the tests), and it works. The stupid fedmsg
consumer hack is untested as yet (but should be tested quite
soon). I'm not married to that, if you hate it we can just drop
it and set up a cron job or something instead.

Test Plan:
Try testing the current live-respins compose, see
if it works. Check I didn't break anything else.

Reviewers: jskladan, garretraziel

Reviewed By: garretraziel

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D1067