Always use dl.fedoraproject.org for downloading ISOs
Forcibly replace download.fp.o with dl.fp.o when setting the job
_URL params. fedfind gives dl.fp.o URLs for most things we test,
but it gives download.fp.o URLs for the post-release respins;
that will frequently give us 404s when they're very new (i.e.
when we want to test them), so just change that. (If we change
things so download.fp.o always resolves to the internal server
in Fedora infra, we can probably live without this).