Improve configurability of consumers and CLI, add tests
Summary:
This is mainly about making the fedmsg consumers configurable.
We also do a bunch of cleaning up in the config settings and
offer more settings in the CLI as well.

I took the opportunity to rework the consumer class naming to
be more consistent; note this will break existing consumer
configs, they will need to be updated. I will update the infra
deployments.

We also add, well, tests. Lots of tests! All the consumers and
the entire results module are tested, which should help us be
confident these changes are OK. I didn't have enough time to
write tests for schedule or cli yet, will do it ASAP.

There's a little bit of general cleanup too - unused imports and
so on.

Test Plan:
Run the tests! There are tests now. Also actually use
the CLI and ideally the consumers and make sure they really do
work, and my tests aren't lies.

Reviewers: jskladan, jsedlak

Reviewed By: jsedlak

Subscribers: tflink

Differential Revision: https://phab.qa.fedoraproject.org/D1118