Unify project name as fedora_openqa
This seems better than setup.py saying 'fedora-openqa', the
library being called `fedora_openqa_schedule` even though it
also does reporting and has a module called `schedule`, and the
whole thing being in a repo called `openqa_fedora_tools`...

The only little problem is that having an underscore in a name
in `setup.py` is fudged a bit behind the scenes, setuptools
changes the name in the egg-info to `fedora-openqa`. But it's
not really a problem, and `fedoraopenqa` seems like an ugly
name to me.