Fix resultsdb reporting for ARM tests (decompressed images)
For the ARM tests, HDD_1 winds up being somefile.raw, but the
original image filename is somefile.raw.xz . We need to pass
the *original* filename to resultsdb_conventions when reporting
the result, so fedfind will be able to find the image. I could
have put some more relaxed image finding logic into fedfind for
this, but I think it's safer just to fix this code to pass the
correct original filename. It's not too hard to handle this
case, if we run into others in future, we can reassess.

I also assigned the `TEST_TARGET` value to a variable instead
of constantly accessing it through the dict, as it makes the
code a bit more readable, I think.