Set `filter_dupes` appropriately when getting openQA jobs
@jsedlak noticed odd results showing up in ResultsDB like this
one:
https://taskotron.fedoraproject.org/resultsdb/results/12481897
Which is a result for job #58646, but shows `NEEDS_INSPECTION`
as the outcome, when the test actually passed. I noticed there
is a later result for the same test:
https://taskotron.fedoraproject.org/resultsdb/results/12482267
with the correct status.

On further investigation, it looks like what happened is this:
58646 is a clone of 58630, which died during execution (causing
58646 to be automatically created). The earlier result is really
supposed to be the result for 58630, not 58646. The consumer saw
that 58630 had completed and called `resultsdb_report` for job
58630. `resultsdb_report` then tries to get the job dict from
openQA, using the client library's `get_jobs` method. However,
it does not specify the `filter_dupes` kwarg. The default value
for that kwarg is `True`, and what that means is that the client
library will notice that 58646 is a dupe of 58630, and send the
job dict for 58646 instead of the one for 58630. So the report
function winds up with the job dict for the running job 58646,
not the finished job 58630, and that's why the outcome winds up
as `NEEDS_INSPECTION`.

So let's adjust our `get_jobs` calls to set `filter_dupes`. I
was kinda unsure what the best behaviour is, but decided on
just making it `True` for wiki reporting and `False` for RDB
reporting. This does mean that when a job gets cloned, we will
wind up trying to wiki report the result for the clone while
it's still running, but this doesn't have any harmful effects as
the wiki reporter only reports passes; it'll just do nothing,
then when the clone does complete, it'll fire up again and do
the right thing.