Tag candidate composes as 'important'
Summary:
openQA has a feature letting you mark builds as 'important' by
adding a comment in a specific format to the job group. This has
a couple of implications. Jobs for the build won't be eaten by
the garbage collector (that otherwise deletes most artifacts of
jobs after a certain time). Also, jobs for the build won't be
cancelled if a 'similar' build is scheduled while they're
running. We had a problem today where all the jobs for F26 Alpha
RC4 were cancelled because an F26 nightly landed shortly after
they started, and openQA decided the nightly jobs should
'obsolete' the RC4 jobs. This will preven that happening again.

Test Plan:
Run the test suite, and if you have a pet openQA
instance, you can test this quite easily by just scheduling
some jobs for a candidate compose and, while they're running,
scheduling jobs for a nightly of the same release. The result
should be that both sets of jobs are run (none are cancelled).

Reviewers: jskladan, jsedlak

Reviewed By: jsedlak

Subscribers: tflink

Differential Revision: https://phab.qa.fedoraproject.org/D1181