Update wiki reporting bits for Base matrix changes
I just changed the Base matrix as recently proposed on the
mailing list; it now has a bunch of sections and the column
titles have changed a bit. These are the appropriate changes
for the wiki reporting bits. Right now checkwiki will report
a couple of errors (for ARM results) because it operates against
the 'current' result pages and the 'current' Base result page
was produced from the *old* matrix not the new one. But as soon
as a new event is created, it should pass (assuming I got the
changes right).