Report base_startup and more base_initial_setup results
Summary:
Now the artwork requirements have been split out of these tests,
we can reasonably report results for them from install_default
and install_no_user jobs.

As checkwiki reports, reporting from Everything and Atomic jobs
will fail, but this isn't really a problem. We should probably
have a 'Minimal' column for x86_64 (but then we'd have to invent
a '$SUBVARIANT_OR_MINIMAL$' or '$PACKAGE_SET$' substitution, or
something, to get from 'Everything' to 'Minimal'), and we could
have an Atomic column, but I don't think we really need to block
on either of those things. For now this will get us all the most
important results.

Test Plan:
Use checkwiki (though it's not quite accurate ATM
because this diff is written to the new Base matrix, but the
current Base page is built from the old matrix) and good ol'
eyeball...

Reviewers: jsedlak, jskladan

Reviewed By: jsedlak

Subscribers: tflink

Differential Revision: https://phab.qa.fedoraproject.org/D1189