Set START_AFTER_TEST as well as HDD_1 in update scheduling
This prevents the scheduler from trying to make update tests
depend on a non-existent install_default_upload test. In the
April snapshot we're currently running, openQA actually goes
ahead and schedules the job anyway, though it prints a warning
message. But this was 'fixed' in May; if a dependency cannot
be found, the job will not be scheduled. So we need to make
sure we don't specify a START_AFTER_TEST value when we don't
want one. There's a corresponding commit for the distri which
uses the '+' trick to ensure the value will still be set for
the two tests where we *do* want it to be set even in the update
case.