Don't resultsdb report EXTRA jobs, add NOREPORT
This commit combines a couple of changes related to sometimes
*not* reporting results.

We've long skipped wiki reporting of results when the job's
'build' value ends in 'EXTRA'; this string is added to the
'build' value when an admin uses the CLI to trigger a set of
tests using an updates image. However, when implementing rdb
reporting, we forgot to do the same there; so this commit adds
that.

The commit also adds another magic string that can be put in
the 'build' value to prevent wiki and resultsdb reporting from
happening: 'NOREPORT'. This is so we can manually trigger
throwaway runs and avoid having the results reported - e.g. I
just ran some tests on an F26 release ISO just to check
something; I didn't want those results sent to rdb or the wiki.
Of course, we *could* just use a 'build' value ending in EXTRA
in this case too, but I thought having a separate magic string
to use would be a bit better; it avoids any possibility of
confusion between 'completely manually triggered job where we
decided we don't want to report the results' and 'job triggered
via the CLI which used the extraparams mechanism'.